djhw/requirements.txt
betfair.py==0.2.2
decorator==4.0.10
django-crontab==0.7.1
enum34==1.1.6
inflection==0.3.1
invoke==0.14.0
psycopg2==2.6.2
python-dateutil==2.6.0
requests==2.12.4
schematics==1.1.1
six==1.10.0
